Rename MobileSentrix scraper
